refactor test structure
