+++ crates/uroborosql-fmt/tests/pgcst_normal_cases.rs
@@ -1,8 +1,8 @@
 use std::fs;
 use std::path::Path;
 
-use console::{style, Style};
-use similar::{ChangeTag, TextDiff};
+mod pgcst_util;
+use pgcst_util::print_diff;
 
 #[derive(Debug)]
 struct TestCase {
@@ -42,54 +42,6 @@ impl TestCase {
             sql,
             expected,
         })
-    }
-}
-
-pub struct Line(Option<usize>);
-
-impl std::fmt::Display for Line {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self.0 {
-            None => write!(f, "    "),
-            Some(idx) => write!(f, "{:<4}", idx + 1),
-        }
-    }
-}
-
-pub fn print_diff(expected: impl Into<String>, got: impl Into<String>) {
-    let expected = expected.into();
-    let got = got.into();
-    let diff = TextDiff::from_lines(&expected, &got);
-
-    for (idx, group) in diff.grouped_ops(3).iter().enumerate() {
-        if idx > 0 {
-            println!("{:-^1$}", "-", 80);
-        }
-        for op in group {
-            for change in diff.iter_inline_changes(op) {
-                let (sign, s) = match change.tag() {
-                    ChangeTag::Delete => ("-", Style::new().red()),
-                    ChangeTag::Insert => ("+", Style::new().green()),
-                    ChangeTag::Equal => (" ", Style::new().dim()),
-                };
-                print!(
-                    "{}{} |{}",
-                    style(Line(change.old_index())).dim(),
-                    style(Line(change.new_index())).dim(),
-                    s.apply_to(sign).bold(),
-                );
-                for (emphasized, value) in change.iter_strings_lossy() {
-                    if emphasized {
-                        print!("{}", s.apply_to(value).underlined().on_black());
-                    } else {
-                        print!("{}", s.apply_to(value));
-                    }
-                }
-                if change.missing_newline() {
-                    println!();
-                }
-            }
-        }
     }
 }
